% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_supported_languages.R
\name{list_supported_languages}
\alias{list_supported_languages}
\title{List supported languages for the World Bank API}
\usage{
list_supported_languages()
}
\value{
A tibble with two columns:
\describe{
\item{language_code}{The ISO 639-1 code of the language (e.g., "en" for English).}
\item{language_name}{The full name of the language (e.g., "English").}
}
}
\description{
This function returns a tibble of supported languages for querying the World Bank API.
The supported languages include English, Spanish, French, Arabic, and Chinese.
}
\details{
This function provides a simple reference for the supported languages when querying
the World Bank API.
}
\examples{
# List all supported languages
list_supported_languages()

}
